# [TraceQL] Add trace level intrinsics