--- a/pkg/search/util.go
+++ b/pkg/search/util.go
@@ -57,5 +57,8 @@ func GetVirtualIntrinsicValues() []string {
 		traceql.IntrinsicKind.String(),
 		traceql.IntrinsicName.String(),
 		traceql.IntrinsicStatus.String(),
+		traceql.IntrinsicTraceDuration.String(),
+		traceql.IntrinsicTraceRootService.String(),
+		traceql.IntrinsicTraceRootSpan.String(),
 	}
 }