--- a/pkg/traceql/engine.go
+++ b/pkg/traceql/engine.go
@@ -53,20 +53,8 @@ func (e *Engine) ExecuteSearch(ctx context.Context, searchReq *tempopb.SearchReq
 	span.SetTag("pipeline", rootExpr.Pipeline)
 	span.SetTag("fetchSpansRequest", fetchSpansRequest)
 
-	// calculate search meta conditions. the only choice is whether or not to include duration
-	// if we request duration as part of the normal span fetch then we can ignore it
-	durationRequested := false
-	for _, c := range fetchSpansRequest.Conditions {
-		if c.Attribute.Intrinsic == IntrinsicDuration {
-			durationRequested = true
-			break
-		}
-	}
-
-	metaConditions := SearchMetaConditions()
-	if durationRequested {
-		metaConditions = SearchMetaConditionsWithoutDuration()
-	}
+	// calculate search meta conditions.
+	metaConditions := SearchMetaConditionsWithout(fetchSpansRequest.Conditions)
 
 	spansetsEvaluated := 0
 	// set up the expression evaluation as a filter to reduce data pulled