--- a/tempodb/encoding/vparquet2/block_traceql.go
+++ b/tempodb/encoding/vparquet2/block_traceql.go
@@ -1648,13 +1648,11 @@ func (c *traceCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 			finalSpanset.Spans = append(finalSpanset.Spans, spanset.Spans...)
 
 			// loop over all spans and add the trace-level attributes
-			if traceAttrs != nil {
-				for k, v := range traceAttrs {
-					for _, sp := range spanset.Spans {
-						s := sp.(*span)
-						if _, alreadyExists := s.attributes[k]; !alreadyExists {
-							s.attributes[k] = v
-						}
+			for k, v := range traceAttrs {
+				for _, sp := range spanset.Spans {
+					s := sp.(*span)
+					if _, alreadyExists := s.attributes[k]; !alreadyExists {
+						s.attributes[k] = v
 					}
 				}
 			}