… CHANGELOG.md
… docs/sources/tempo/traceql/_index.md
… modules/querier/querier_test.go
… pkg/traceql/expr.y.go
… pkg/traceql/lexer.go
… pkg/traceql/parse_test.go
… pkg/traceql/storage_test.go
… pkg/traceql/test_examples.yaml
… tempodb/encoding/vparquet/block_search_tags_test.go
… tempodb/encoding/vparquet/block_traceql_meta_test.go
… tempodb/encoding/vparquet2/block_search_tags_test.go
… tempodb/encoding/vparquet2/block_traceql_meta_test.go
… tempodb/tempodb_search_test.go