**What this PR does**:
This PR adds trace level intrinsics: `traceDuration`, `rootServiceName` and `rootName`.

**Which issue(s) this PR fixes**:
This is not quite the same as adding a scope but it does address the major needs in #1989 

**Checklist**
- [x] Tests updated
- [x] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`